align RPC protocol data, separating it into its own module
Most of this code was already present, but resided in `Kore.JsonRpc`.
The separate module makes it easier to re-use the code elsewhere.

New: `stop-reason: aborted` (never returned by code in this repository)

Also added some additional `FromJSON` and `ToJSON` instances
