+++ kore/kore.cabal
@@ -367,6 +367,7 @@ library
     Kore.Internal.TermLike.TermLike
     Kore.Internal.Variable
     Kore.JsonRpc
+    Kore.JsonRpc.Base
     Kore.Log
     Kore.Log.DebugAppliedRewriteRules
     Kore.Log.DebugAttemptedRewriteRules
